+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix: Match numerical and string IDs properly in `compose(match_by='XYZ')`
 - Fix: Ensure that the domain and histograms match by avoiding missing categorical indices
 - Fix: Ignore `NaN`s when computing histograms
+- Fix: Warn when data contains `NaN`s and replace them with zeros
 
 ## v0.16.0